requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.9
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.0
pipreqs==0.5.0
platformdirs==4.3.3
playwright==1.47.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
pyee==12.0.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
psycopg2-binary==2.9.9
requests==2.32.3
rpds-py==0.20.0
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.11.1
